Fixed Screen Compatibility Issues
I changed the specifications of the screen setup. It should work nicer with most displays now. I also previously added a bug by setting the default resolution to 1980x1080, so I fixed that by changing it to the intended 1920x1080, and then redid the dynamic screen size to work nicer with that default dimension set.